requirements.txt
selenium>=4.0.0
requests>=2.25.1
webdriver-manager>=3.5.0
beautifulsoup4>=4.9.3
Pillow>=9.0.0
